API (Exported Functions)

RadioTransients.Filterbank

RadioTransients.Filterbank

RadioTransients.Freq

RadioTransients.dedisperse

RadioTransients.estimate_dm

RadioTransients.Δt

RadioTransients.dedisperse — Method

dedisperse(dyn_spec, DM)

Creates a de-dispersed dynamic spectrum for dispersion measure DM in pc/cc given that the source filterbank dynamicspectrum `dynspechas samples with time stept_step` in seconds.

RadioTransients.estimate_dm — Method

estimatedm(dynspec)

Estimates the dispersion measure for filterbank dynamicspectrum `dynspec` by maximizing the standard deviation of the folded spectrum.

RadioTransients.Δt — Method

Δt(DM, ν₁, ν₂)

Calculates the time delay corresponding to a dispersed pulse at dispersion measure DM in pc/cc between frequencies ν₁ and ν₂

RadioTransients.Filterbank — Type

Filterbank(data,headers)

The datastructure that holds the the Filterbank file.

Fields

data: The data as a DimArray with dimensions Samp and Freq

headers: The dictionary of header information

RadioTransients.Filterbank — Method

Filterbank("file.fil")

Read a SIGPROC .fil file into a Filterbank.

Arguments

filename::String: The name of the file

Optional Arguments

start::Int: The starting time sample to read from (inclusive)

stop::Int: The stopping time sample to read to (inclusive)

header_int::DataType: The type of header integers, defaults to UInt32

header_float::DataType: The type of header flots, defaults to Float64

RadioTransients.Freq — Type

The frequency dimension of Filterbank data in MHz