API (Exported Functions)

SIGPROC.Filterbank

SIGPROC.Filterbank

SIGPROC.Freq

SIGPROC.fake_pulse

SIGPROC.fake_pulse — Method

fakepulse(DM,flow,f_high)

Generates a Filterbank file corresponding to a fake pulse of dispersion measure DM from frequencies f_low to f_high (MHz).

Optional Arguments

channels: Number of frequency channels

samples: Number of time samples

'δt`: The time step (s) represented by one sample

w: Width of the pulse in time samples

A: Amplitude of pulse in SNR

α: Spectral index, i.e. attenuation of pulse over frequency

start: Index of start of pulse

noise_floor: Value of the top of the noise floor. This is 1 for dtype of Floats and one tenth typemax for integers by default.

dtype: Data type of data

SIGPROC.Filterbank — Type

Filterbank(data,headers)

The datastructure that holds the the Filterbank file.

Fields

data: The data as a DimArray with dimensions Samp and Freq

headers: The dictionary of header information

SIGPROC.Filterbank — Method

Filterbank("file.fil")

Read a SIGPROC .fil file into a Filterbank.

Arguments

filename::String: The name of the file

Optional Arguments

start::Int: The starting time sample to read from (inclusive)

stop::Int: The stopping time sample to read to (inclusive)

header_int::DataType: The type of header integers, defaults to UInt32

header_float::DataType: The type of header flots, defaults to Float64

SIGPROC.Freq — Type

The frequency dimension of Filterbank data in MHz